requirements.txt
sentence-transformers==2.2.2
pandas
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
from_root
boto3
mypy-boto3-s3
botocore
fastapi
uvicorn
jinja2
-e .